requirements.txt
python-dotenv
APScheduler==3.11.0
discord.py==2.6.3